+++ CHANGELOG.md
@@ -2,12 +2,19 @@
 
 ## [Unreleased][]
 
-[Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-aws/compare/0.8.0...HEAD
+[Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-aws/compare/0.9.0...HEAD
+
+## [0.9.0][]
+
+[0.9.0]: https://github.com/chaostoolkit-incubator/chaostoolkit-aws/compare/0.8.0...0.9.0
 
 ### Added
 
 -   support for Python 3.7
 -   support for cloudwatch probes and actions
+-   support for invoking lambdas
+-   support for getting and updating lambda timeout and memory size limits
+-   probe for cloudwatch metric statistics
 
 ## [0.8.0][]